% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_ar_l2.R
\name{cv_ar_l2}
\alias{cv_ar_l2}
\title{Cross-validation for a L2 autoregression model}
\usage{
cv_ar_l2(x, folds = 5, lags = 7, beta = 0.1)
}
\arguments{
\item{x}{vector, time series to be used for cross validation}

\item{folds}{integer, number of folds considered for CV}

\item{lags}{interger, number of lags}

\item{beta}{float, reg. param. for regularization param}
}
\value{
List with,
\itemize{
\item fold_mae: vector, each validation MAE per fold
\item mean_mae: float, mean MAE across folds
\item lags: integer, number of lags used
\item beta: float, regularization factor
}
}
\description{
Cross-validation for a L2 autoregression model
}
\examples{
output <- cv_ar_l2(1:365, folds = 12, lags = 7, beta = 0.1)
print(output$fold_mae)
print(output$mean_mae)
print(output$integer)
print(output$beta)
}
