% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{cv_dataset_creation}
\alias{cv_dataset_creation}
\title{Title}
\usage{
cv_dataset_creation(x, folds = 5)
}
\arguments{
\item{x}{vector, time series data}

\item{folds}{integer, number of folds for CV}
}
\value{
list, with folds for training and validation
}
\description{
Title
}
\examples{
result <- cv_dataset_creation(1:25, 5)
}
