% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_regression.R
\name{dynamic_regression}
\alias{dynamic_regression}
\title{Dynamic training/testing}
\usage{
dynamic_regression(x_train, x_test, beta, lags)
}
\arguments{
\item{x_train}{training univariate time serie}

\item{x_test}{testing univariate time serie}

\item{beta}{optimized regularization param}

\item{lags}{optimized number of lags}
}
\value{
\itemize{
\item y_pred: vector, out-of-sample predictions
\item lags: integer, number of used lags
\item beta: float, value of used regularization param
}
}
\description{
Dynamic training/testing
}
\examples{
x <- seq(1,2, length.out = 100)
x_train <- x[1:80]
x_test <- x[81:100]
beta <- 1e-5
lags <- 1
out <- dynamic_regression(x_train, x_test, beta, lags)
print(x_test)
print(out$y_pred)
}
