% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmse.R
\name{rmse}
\alias{rmse}
\title{Root mean square error (RMSE)}
\usage{
rmse(y_gt, y_pred)
}
\arguments{
\item{y_gt}{numeric, ground-truth time serie}

\item{y_pred}{numeric, prediction time serie}
}
\value{
scalar, RMSE value
}
\description{
Computes the RMSE between one ground-truth time serie and a respective time series
}
\examples{
y_gt <- 1:15
y_pred <- y_gt + rnorm(15, 0, 0.1)
rmse_value <- rmse(y_gt, y_pred)
print(rmse_value)
}
