% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soft.R
\name{soft}
\alias{soft}
\title{Soft-thresholding}
\usage{
soft(a, lambda)
}
\arguments{
\item{a}{the first argument to soft-thresholding function; numeric scalar}

\item{lambda}{thresholding parameter; positive, numeric scalar}
}
\value{
a numeric scalar
}
\description{
Calculates soft-thresholding function
}
\examples{
soft(5, 3)
}
