% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_austin-data.R
\docType{data}
\name{temp_austin}
\alias{temp_austin}
\title{DAily temperature data for Austin, Texas}
\format{
A data frame with 3557 rows and the daily temperature measurement for Austin (on average humidity conditions):
\describe{
\item{X}{index}
\item{date}{date, YYYY-MM-DD}
\item{temp}{numeric, Temperature on average humidity conditions (°F)}
\item{year}{numeric, Correspondent year}
}
}
\source{
\url{https://www.ncei.noaa.gov/} (NCEI / NOAA Climate Data Online)
}
\usage{
temp_austin
}
\description{
A dataset containing daily weather observations for Austin, Texas.
}
\keyword{datasets}
