% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_split.R
\name{train_test_split}
\alias{train_test_split}
\title{Split function to obtain train and test datasets.}
\usage{
train_test_split(x, dates, p = 0.05)
}
\arguments{
\item{x}{vector, the total univariate time series to be considered.}

\item{dates}{vector, the respective dates for x (same length as x).}

\item{p}{numeric, percentage of the time series used for testing,
default = 0.05 (5\\%).}
}
\value{
List with:
\itemize{
\item x_train: vector, training set
\item x_test: vector, testing set
\item dates_train: vector, training dates
\item dates_test: vector, testing dates
}
}
\description{
Split function to obtain train and test datasets.
}
\examples{
x <- 1:20
dates <- as.Date("2024-01-01") + 0:19
split <- train_test_split(x, dates, p = 0.2)
split$x_train
split$x_test
split$dates_train
split$dates_test
}
