% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynamic_regression.R
\name{dynamic_regression}
\alias{dynamic_regression}
\title{Dynamic ridge autoregressive forecasting}
\usage{
dynamic_regression(x_train, x_test, beta, lags, window)
}
\arguments{
\item{x_train}{numeric vector, training time series.}

\item{x_test}{numeric vector, test time series (future observed values).}

\item{beta}{non–negative scalar, ridge regularization parameter (lambda).}

\item{lags}{integer, number of lags in the AR(lags) model.}

\item{window}{integer, size of the rolling training window at each step.
Must satisfy \code{window > lags} and \code{window <= length(x_train)}.}
}
\value{
numeric vector of predictions of length \code{length(x_test)}.
}
\description{
This function performs a dynamic training scheme of a ridge-autoregressive model over
a univariate time-series. At each step it:
\itemize{
\item fits an AR(lags) model (where lags = # of lags) with ridge penalty using only the last \code{window}
observations from the training part of the current time series,
\item predicts a one-step-ahead forecast,
\item updates the training series by dropping its first value and
appending the next observed value from \code{x_test}, maintaining
the length of the training batch equal to the \code{window}.
}
The number of predictions is equal to \code{length(x_test)}.
}
\details{
The model includes an intercept which is not penalized.
}
\examples{
set.seed(2025)
x <- arima.sim(model = list(ar = 0.7), n = 200)
x_train <- x[1:160]
x_test  <- x[161:200]
preds <- dynamic_regression(x_train, x_test,
                          beta = 1, lags = 3, window = 50)
length(preds)  # should be length(x_test)
}
